pkb_json/requirements.txt
Django==1.9.4
django-cors-headers==1.1.0
django-jsonview==0.5.0
goto==0.1.3
MySQL-python==1.2.5
PyJWT==1.4.0
